% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncommon.R
\name{\%<=\%}
\alias{\%<=\%}
\title{Multiple assignment operator}
\usage{
x \%<=\% value
}
\arguments{
\item{x}{A name structure, see section below.}

\item{value}{A list of values, vector of values, or \R object to assign.}
}
\description{
Avoid conflict with future::`\%<-\%`.
}
