% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncommon.R
\docType{import}
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for `NULL`}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang]{\%||\%}}}
}}

